+++ test_project/fake_app/admin.py
@@ -14,6 +14,22 @@ class ArtistAdmin(CeleryImportExportMixin, admin.ModelAdmin):
         "name",
         "instrument",
     )
+    list_filter = (
+        "instrument__title",
+    )
+    search_fields = (
+        "instrument__title",
+        "name",
+        # `@name`, `^name`, `=name` are about one model field but thanks for
+        # first symbol they apply difference lookups in search.
+        # These ones are used to check that type of search filter
+        # must be only one for a model field in admin class.
+        # They are not required, it's just for testing
+        # https://github.com/django/django/blob/d6925f0d6beb3c08ae24bdb8fd83ddb13d1756e4/django/contrib/admin/options.py#L1138
+        "@name",
+        "^name",
+        "=name",
+    )
     resource_classes = [ArtistResourceWithM2M, SimpleArtistResource]